auth/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
jinja2==3.1.3
